subpictures: Allow vout plugins to do native scaling
Subpictures that cannot be rendered in output size directly (all kind of
picture subtitles) have been scaled to output size using swscale so far. For
those output plugins that can blend the subpictures directly to the video it
makes sense to allow offloading of the scaling to the output plugins as well,
as they can likely scale hardware accelerated.
The positioning information is scaled as before, just the actual picture size
is untouched. Additionally the region structure holds a new pair of size
values, which informs the video output about the actual output size which the
picture has to be scaled to.

Signed-off-by: Julian Scheel <julian@jusst.de>