subpictures: Use correct rounding for scaled sizes
Especially when subpictures consist of several region pictures that are placed
right next to each other (like seen with dvb subtitles) it is important that
the positioning of the regions is exact. To achieve this correctly round the
scaled values instead of just casting to int.
This fixes several rendering glitches with scaled subtitles.

Signed-off-by: Julian Scheel <julian@jusst.de>