h264 packetizer: No infinite wait for SEI recovery information
The SEI recovery point information is not mandatory in the h264 specification.
If it is included it has to be transmitted before each primary coded picture.
So to ensure that we see use it if present but don't wait too long if not
present we add a picture count. If the picture count reaches 2 and we saw no
SEI information there won't be any in the stream, so we leave the PREROLL
state.

Signed-off-by: Julian Scheel <julian@jusst.de>