vout: Allow vout plugins to request phase offset
If the display refresh rate matches the video rate it is desired to show
exactly one picture per frame. Some video renderers take care of this
internally others don't. For those that don't do it we allow the vout to
request a phase offset which is applied to the picture display timestamps, so
that the vout can ensure that pictures are rendered in the desired phase shift
to the display refresh.

Signed-off-by: Julian Scheel <julian@jusst.de>