no-upstream-yet: mmal/vout: Request explicit phase offset at core
Applying the phase offset in vd_prepare does not work well because VLCs timing
mechanism which decides whether there is enough time to update subpictures,
etc. is using picture dates before prepare is called.
For this we introduced a mechanism in the core that allows us to specify a
phase offset to be applied by the core. This applies the offset right after
fetching the picture from the decoder queue, which makes sure that the timing
mechanisms still work.

Signed-off-by: Julian Scheel <julian@jusst.de>